update test_server to support uploader function
