+++ yandex_types.py
@@ -1,9 +1,9 @@
-from typing import TypedDict, Dict
+from typing import TypedDict, Dict, Union, Mapping
 
 
 class YandexEvent(TypedDict):
     httpMethod: str
-    headers: Dict[str, str]
-    body: str
+    headers: Mapping[str, str]
+    body: Union[str, bytes]
     queryStringParameters: Dict[str, str]
     isBase64Encoded: bool
